# Optimize queries for ERC20 balances