@JagoFigueroa 2024-09-23
Looking good to me, haven't seen any issues on staging and was a nice opportunity to try some performance testing too!

I have let the wallet team QAs know of this optimisation too so that they can keep an eye in case they see something weird while testing. 

Maybe before deploying this to prod we can sync and execute a performance test run there of the before and after and compare the results too 😄 

Cheers!